// types/home.ts
export interface NowPlayingParams {
  language?: string;
  api_key?: string;
  page?: number;
}

export interface DefaultResponse {
  status_code?: number;
  status_message?: string;
  success?: boolean;
}

export interface Dates {
  maximum?: string;
  minimum?: string;
}

export interface NowPlayingResponse {
  status_code?: number;
  status_message?: string;
  success?: boolean;
  dates: Dates;
  page: number;
  results: DataNowPlaying[];
  total_pages: number;
  total_results: number;
}

export interface DataNowPlaying {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface Genre {
  id: number;
  name: string;
}

export interface MovieDetailResponse {
  adult: boolean;
  backdrop_path: string;
  belongs_to_collection: BelongsToCollection;
  budget: number;
  genres: Genre[];
  homepage: string;
  id: number;
  imdb_id: string;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  production_companies: ProductionCompany[];
  production_countries: ProductionCountry[];
  release_date: string;
  revenue: number;
  runtime: number;
  spoken_languages: SpokenLanguage[];
  status: string;
  tagline: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface ProductionCompany {
  id: number;
  logo_path: any;
  name: string;
  origin_country: string;
}

export interface ProductionCountry {
  iso_3166_1: string;
  name: string;
}

export interface SpokenLanguage {
  english_name: string;
  iso_639_1: string;
  name: string;
}

export interface BelongsToCollection {
  id: number;
  name: string;
  poster_path: string;
  backdrop_path: string;
}

export interface MovieDetailParams {
  movie_id: number;
  language?: string;
  api_key?: string;
}

export interface CharacterDetailParams {
  person_id: number;
  language?: string;
  api_key?: string;
}

export interface Character {
  id: number;
  cast: Cast[];
  crew: Crew[];
}

export interface Cast {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path?: string;
  cast_id: number;
  character: string;
  credit_id: string;
  order: number;
}

export interface Crew {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path?: string;
  credit_id: string;
  department: string;
  job: string;
}

export interface VideoTrailer {
  id: number;
  results: Result[];
}

export interface Result {
  iso_639_1: string;
  iso_3166_1: string;
  name: string;
  key: string;
  site: string;
  size: number;
  type: string;
  official: boolean;
  published_at: string;
  id: string;
}

export interface CharacterDetail {
  adult: boolean;
  also_known_as: string[];
  biography: string;
  birthday: string;
  deathday: any;
  gender: number;
  homepage: any;
  id: number;
  imdb_id: string;
  known_for_department: string;
  name: string;
  place_of_birth: string;
  popularity: number;
  profile_path: string;
}

export interface MovieCredits {
  cast: CastCredits[];
  crew: CrewCredits[];
  id: number;
}

export interface CastCredits {
  adult: boolean;
  backdrop_path?: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path?: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
  character: string;
  credit_id: string;
  order: number;
}

export interface CrewCredits {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
  credit_id: string;
  department: string;
  job: string;
}

export interface HomeServiceType {
  fetchListNowPlaying(
    params: NowPlayingParams,
  ): AxiosPromise<NowPlayingResponse>;
  fetchListPopular(params: NowPlayingParams): AxiosPromise<NowPlayingResponse>;
  fetchListTopRated(params: NowPlayingParams): AxiosPromise<NowPlayingResponse>;
  fetchMovieDetail(
    params: MovieDetailParams,
  ): AxiosPromise<MovieDetailResponse>;
  fetchCharacter(params: MovieDetailParams): AxiosPromise<Character>;
  fetchMovieTrailer(params: MovieDetailParams): AxiosPromise<VideoTrailer>;
  fetchCharacterDetail(
    params: CharacterDetailParams,
  ): AxiosPromise<CharacterDetail>;
  fetchMovieCredits(params: CharacterDetailParams): AxiosPromise<MovieCredits>;
}